+++ src/API/JobProcessorBase.cs
@@ -1,6 +1,6 @@
 ﻿/*
  * Apache License, Version 2.0
- * Copyright 2019-2020 NVIDIA Corporation
+ * Copyright 2019-2021 NVIDIA Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -88,7 +88,10 @@ namespace Nvidia.Clara.DicomAdapter.API
 
             _logger.Log(LogLevel.Information, "Queueing a new job '{0}' with pipeline '{1}', priority={2}, instance count={3}", jobName, pipelineId, jobPriority, instances.Count);
 
-            var job = await _jobsApi.Create(pipelineId, jobName, jobPriority);
+            var metadata = new JobMetadataBuilder();
+            metadata.AddSource($"{Name}-{AeTitle}");
+
+            var job = await _jobsApi.Create(pipelineId, jobName, jobPriority, metadata);
             using (_logger.BeginScope(new LogginDataDictionary<string, object> { { "JobId", job.JobId }, { "PayloadId", job.PayloadId } }))
             {
                 await _jobStore.Add(job, jobName, instances);