+++ src/Server/Services/Http/InferenceController.cs
@@ -141,7 +141,11 @@ namespace Nvidia.Clara.DicomAdapter.Server.Services.Http
 
         private async Task CreateJob(InferenceRequest request)
         {
-            var job = await _jobsApi.Create(request.Algorithm.PipelineId, request.JobName, request.ClaraJobPriority);
+            
+            var metadata = new JobMetadataBuilder();
+            metadata.AddSource($"{request.TransactionId}");
+
+            var job = await _jobsApi.Create(request.Algorithm.PipelineId, request.JobName, request.ClaraJobPriority, metadata);
             request.JobId = job.JobId;
             request.PayloadId = job.PayloadId;
         }